Improve test environment
Now there are two run configurations that you can use in PyCharm to
run all tests: either in the IDE, or from the IDE in the shell.
Note that all test file names should match 'test_*.py' for module
unittest to discover the file and run it. Travis now also is more
verbose.