Improve adding of vertices to a graph
Preventing duplicates is required for the complement method coming in a
future PR. The complement needs to add many edges and in the process
many vertices were added multiple times. This commit prevents that.

The list of vertices of a graph could also be refactored to a set.
The list of graphs of a vertex could also be refactored to a set.