Improve graph.py and test environment (#33)
* Reformat graph.py

* Improve test environment

Now there are two run configurations that you can use in PyCharm to
run all tests: either in the IDE, or from the IDE in the shell.
Note that all test file names should match 'test_*.py' for module
unittest to discover the file and run it. Travis now also is more
verbose.

* Clean up graph.py

* Add unique integer generation tool

* Add name property setter to Graph class

* Improve graph addition (disjoint union)

* Fix doc

* Improve graph disjoint union

* Remove local Python interpreter dependency 2/2

* Add Graph class properties order and size

* Add Graph class label property

* Remove unused import

* Add documentation newlines

* Remove not applicable comment

* Refactor compare to accept Iterable type

* Improve assertion

* Improve printing of Graph objects

* Improve adding of vertices to a graph

Preventing duplicates is required for the complement method coming in a
future PR. The complement needs to add many edges and in the process
many vertices were added multiple times. This commit prevents that.

The list of vertices of a graph could also be refactored to a set.
The list of graphs of a vertex could also be refactored to a set.

* Fix run config 3/2 (rebasing failed somehow)

* Adds test for adding the same graph to eachother.

* Revert "Improve adding of vertices to a graph"

This reverts commit 446c96b1ce462acfa12ef8f190d22de0362e48c9.

* Fix disjoint union operation

* Fix doc

* Use variable name instead of variable definition

* Use create_graph_helper