Keep track of known isomorphisms (#53)
* Keep track of known (an)isomorphisms

* Fix storing of known (an)isomorphisms

Known anisomorphisms don't propagate knowledge to other known
anisomorphisms. Possibly, in a future commit the known anisomorphisms
will be removed entirely to simplify things further.

* Reformat test_fast_color_refine.py

* Add integration test for known (an)isomorphisms

* Remove keeping track of known anisomorphisms

* Rename test method

* Add triple isomorphisms to test graphs

The previous situation tested with two isomorphisms of a graph that
only has two. The current situation tests 3 of 6 possible isomorphisms.

* Return known isomorphisms form main method

This allows for a much better module design, it cleans up a test file
and it makes testing a lot easier.